+++ src/lib/command_test.rs
@@ -1,7 +1,6 @@
 use super::*;
 use crate::test;
 use crate::types::Task;
-use ci_info;
 use std::io::ErrorKind;
 
 #[test]
@@ -54,13 +53,13 @@ fn is_silent_for_level_other() {
 #[test]
 fn should_print_commands_for_level_error() {
     let print_commands = should_print_commands_for_level("error".to_string());
-    assert_eq!(print_commands, ci_info::is_ci());
+    assert!(!print_commands)
 }
 
 #[test]
 fn should_print_commands_for_level_info() {
     let print_commands = should_print_commands_for_level("info".to_string());
-    assert_eq!(print_commands, ci_info::is_ci());
+    assert!(!print_commands)
 }
 
 #[test]
@@ -72,7 +71,7 @@ fn should_print_commands_for_level_verbose() {
 #[test]
 fn should_print_commands_for_level_other() {
     let print_commands = should_print_commands_for_level("test".to_string());
-    assert_eq!(print_commands, ci_info::is_ci());
+    assert!(!print_commands)
 }
 
 #[test]